backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy[asyncio]>=2.0.36
asyncpg>=0.29.0
alembic>=1.13.3
pydantic>=2.9.0
pydantic-settings>=2.5.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.12
celery>=5.4.0
redis>=5.1.0
anthropic>=0.40.0
httpx>=0.27.0
supabase>=2.10.0
python-dotenv>=1.0.0
aiofiles>=24.1.0
duckdb>=1.1.0
pandas>=2.2.0
pyarrow>=17.0.0
openpyxl>=3.1.0
python-slugify>=8.0.0
